Remove CaseloadDetailView
