+++ visitz/Visitz/Views/Caseload/CaseloadContainerView.xaml
@@ -1,10 +1,13 @@
 <?xml version="1.0" encoding="utf-8" ?>
-<split:SplitLayoutView
+<base:BaseContentView
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:base="clr-namespace:Visitz.Views.BaseClasses"
     xmlns:caseload="clr-namespace:Visitz.Views.Caseload"
-    xmlns:split="clr-namespace:Visitz.Views.SplitView"
+    xmlns:nav="clr-namespace:Visitz.Views.Navigation"
     x:Class="Visitz.Views.Caseload.CaseloadContainerView"
     x:DataType="caseload:CaseloadContainerView"
     BackgroundColor="White"
-></split:SplitLayoutView>
+>
+    <nav:ContentViewNavigationStack x:Name="ContentStack" />
+</base:BaseContentView>
