--- visitz/Visitz/Views/Caseload/CaseloadDetailView.xaml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<baseClasses:ViewModelContentView
-    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:baseClasses="clr-namespace:Visitz.Views.BaseClasses"
-    xmlns:caseload="clr-namespace:Visitz.Views.Caseload"
-    xmlns:views="clr-namespace:Visitz.Views"
-    x:Class="Visitz.Views.Caseload.CaseloadDetailView"
-    x:DataType="caseload:CaseloadDetailViewModel"
->
-    <AbsoluteLayout>
-        <views:WatermarkView
-            AbsoluteLayout.LayoutFlags="PositionProportional"
-            AbsoluteLayout.LayoutBounds="0.5,0.5"
-        />
-
-        <Grid AbsoluteLayout.LayoutFlags="All" AbsoluteLayout.LayoutBounds="1,1,1,0.1">
-            <views:LastUpdatedBanner
-                HorizontalOptions="FillAndExpand"
-                VerticalOptions="End"
-                LastUpdated="{Binding LastUpdated}"
-            />
-        </Grid>
-    </AbsoluteLayout>
-</baseClasses:ViewModelContentView>
