+++ visitz/Visitz/Views/Caseload/CaseloadViewModel.cs
@@ -14,10 +14,12 @@ using Visitz.Services.Caseload;
 using Visitz.Storage;
 using Visitz.Views.BaseClasses;
 using Visitz.Views.SegmentedButtons;
+using VisitzModel.Events;
 using VisitzModel.Extensions;
 using VisitzModel.Messaging;
 using VisitzModel.Models.Caseload;
 using VisitzModel.Models.EntityTypes;
+using VisitzModel.Storage;
 using IBusinessObjectExtensions = VisitzModel.Models.Caseload.IBusinessObjectExtensions;
 
 namespace Visitz.Views.Caseload
@@ -111,6 +113,11 @@ namespace Visitz.Views.Caseload
 
         OidcSessionInfo SessionInfo { get; set; }
 
+        LastUpdatedPrefs LastUpdatedPrefs { get; set; } = ServiceProvider.GetService<LastUpdatedPrefs>();
+
+        [ObservableProperty]
+        public DateTime? lastUpdated;
+
         private async Task Setup()
         {
             WeakReferenceMessenger.Default.Register(this, GetAllDataForOfflineService.MakeId());
@@ -133,6 +140,9 @@ namespace Visitz.Views.Caseload
 
             DeviceDisplay.Current.MainDisplayInfoChanged += Current_MainDisplayInfoChanged;
             ShowMenuButton = DeviceDisplay.Current.MainDisplayInfo.Orientation == DisplayOrientation.Portrait;
+
+            LastUpdated = LastUpdatedPrefs.Get(GetCaseloadService.MakeId());
+            LastUpdatedPrefs.LastUpdatedChanged += LastUpdatedPrefs_LastUpdatedChanged;
 
             WeakReferenceMessenger.Default.Send(AutoRefreshService.MakeStartMessage());
         }
@@ -402,6 +412,12 @@ namespace Visitz.Views.Caseload
         void ReceiveNavBarPositionMessage(object recipient, NavPositionMessage message)
         {
             ShowMenuButton = ((TwoPaneViewMode)message.Value) == TwoPaneViewMode.Tall;
+        }
+
+        private void LastUpdatedPrefs_LastUpdatedChanged(object sender, LastUpdatedChangedEventArgs e)
+        {
+            if (e.Id.Equals(GetCaseloadService.MakeId()))
+                LastUpdated = (sender as LastUpdatedPrefs).Get(e.Id);
         }
     }
 }
