+++ visitz/Visitz/Views/Navigation/ContentViewNavigationStack.cs
@@ -6,9 +6,20 @@ internal partial class ContentViewNavigationStack : ContentView
 {
     readonly Stack<ContentView> _viewStack = new();
 
+    readonly AbsoluteLayout _layout = [];
+
+    public ContentViewNavigationStack()
+    {
+        Content = _layout;
+    }
+
     public async Task PushAsync(ContentView newView)
     {
         _viewStack.Push(newView);
+        _layout.Add(newView);
+
+        AbsoluteLayout.SetLayoutBounds(newView, new Rect(0, 0, 1, 1));
+        AbsoluteLayout.SetLayoutFlags(newView, Microsoft.Maui.Layouts.AbsoluteLayoutFlags.All);
 
         newView.TranslationX = X + Width;
 
@@ -21,6 +32,9 @@ internal partial class ContentViewNavigationStack : ContentView
 
         await view.TranslateToAsync(X + view.Width, Y, easing: Easing.CubicInOut);
 
-        return _viewStack.Pop();
+        _viewStack.Pop();
+        _layout.Remove(view);
+
+        return view;
     }
 }
