--- visitz/Visitz/Views/WatermarkView.xaml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<ContentView
-    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:views="clr-namespace:Visitz.Views"
-    x:Class="Visitz.Views.WatermarkView"
-    x:DataType="views:WatermarkView"
->
-    <Image
-        HeightRequest="300"
-        WidthRequest="300"
-        Opacity="0.25"
-        HorizontalOptions="Center"
-        VerticalOptions="Center"
-        Source="bc_logo_vertical_positive.png"
-    />
-</ContentView>
