--- visitz/Visitz/Views/WatermarkView.xaml.cs
@@ -1,9 +1,0 @@
-namespace Visitz.Views;
-
-public partial class WatermarkView : ContentView
-{
-    public WatermarkView()
-    {
-        InitializeComponent();
-    }
-}
